Wire up fsi_bit_bang part
The fsi_bit_bang part is used for specifying which individual BMC
GPIOs are used as part of the FSI Bus.  So the BMC connects GPIOs
to it, and then it has an fsi-master unit to connect to the
processor's FSI slave.  This is required so OpenFSI can know which
GPIO is which.